Refactorized server code
