+++ cli/src/main.rs
@@ -1,5 +1,4 @@
 #![feature(proc_macro_hygiene, decl_macro)]
-
 extern crate chrono;
 extern crate crypto;
 extern crate dotenv;
@@ -10,8 +9,10 @@ extern crate rocket_contrib;
 #[macro_use]
 extern crate serde_derive;
 extern crate rnotes_core;
+extern crate rnotes_server;
 
 mod handlers;
+mod utils;
 
 use dotenv::dotenv;
 use rnotes_core::BDPool;
