+++ core/src/models/api/auth.rs
@@ -1,0 +1,10 @@
+#[derive(Serialize, Deserialize)]
+pub struct LoginIn {
+    pub email: String,
+    pub password: String,
+}
+
+#[derive(Serialize, Deserialize)]
+pub struct LoginOut {
+    pub jwt_token: String,
+}
