+++ core/src/models/api/mod.rs
@@ -1,0 +1,3 @@
+pub mod auth;
+pub mod category;
+pub mod note;
