+++ core/src/models/db/mod.rs
@@ -1,2 +1,3 @@
 pub mod category;
 pub mod note;
+pub mod user;
