+++ server/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rnotes_server"
-version = "0.1.0"
+version = "0.3.0"
 authors = ["Alvaro Santos Andres <alvsanand@gmail.com>"]
 edition = "2018"
